Test 5: testDynamicEval
Could not find any value/variable to Parameterize in test 2, 3 & 4. Found 2 hard coded values in test 5: (1) String key and (2) boolean value of the context to be set. Replaced multiple occurances of both in set methods and assertions with the introduced parameterized variables. Added 5 value sets, the test ran successfull for all.