fixing test 8
For 4th value set :-  2nd assertion inside method assertContextValue failed with error: Expecting not blank but was:<“”>. Inside the method it explicitly asserts for the value (which was set to key itself before) not to be a blank string. Hence, this input was bad and can be removed. We can update the test to take key and value separately to test key as an empty string, but it was already covered in test 2.   For 5th value set:-   This time 1st assertion failed inside method assertContextValueIsNull with error:- Expecting value to be equal to null but was not. This happened as (1) the test treats both blank string and string with only spaces as same and (2) in the previous run (for value set 4) the context for blank string was set but was never removed (which happens after assertion 2). Hence it found an empty object in its value which was supposed to be null. On running this value set individually it fails just like value set 4, suggesting it a bad value that can we removed. Removed both the bad value sets. The whole test passes successfully.