Test 1 :- testCompressorPoolCounts
Found 2 parametes :- (1) Number of compressors to get and check count, created an array to store and return compressors -> created 2 for loops to assert while getting/returning compressors. (2) No of times to check if 0 is returned when getLeasedCompressorsCount is called with no compressors to return, also created for loop for this. Passes value set as (2, 2) default values from test and the test ran success fully.