Test 4: testMultiThreadedCompressorPool
Found and added 2 parameters:- (1) local variable iterations -> removed declaration and initialization. (2) hard-coded value of thread count -> replaced with param. Added a total of 8 value sets out of which only 1 and 3 passesed. 2 & 4 failed due to timeout -> as we provided only 1 thread which would cause a dead lock due to lack of resources as the 1st comsumer will keep on waiting to consume and never leave the thread, and the producer will never get the thread to produce. 5, 6, 7 & 8 failed as neither the thread pool nor LinkedBlockingDeque can be initialized with 0 or -ve values. Hence, all failures are for bad values.