fixed and imporved Test 4
fixed bad values (replaced 1 for nThreads with 2) and removed 0/-ve bad values. Also found the queue was being initialized with 2 * iterations, which would never be needed as only Producer pushes in queue and the consumer only de-queues so the max size of the queue needed will be equal to number of produces i.e iterations. Removed multiply by 2 in queue size. All tests ran successfully.