fixed by adding assume statement
The error was caused in a assertion statement of checkBlockLocation() in line 152, as we passed a null in array value, assertArrayEquals failed as it cannot have an expected value as null.
The test already checks for null by setting them, it works as checkBlockLocation() in line 141 does have assertArrayEquals but it passes an empty array by default.
To fix, added assume statement to check if any of array value is null.
Question:- In this case should be consider splitting the test as null check is being redundantly checked for all value sets.