Bug Fixed and added assume statement
For all -ve values of dataSizeMB which passes -> We found there were not explicit checks to only take +ve values of data size, given we have such a check (indirect though) for chunkSizeKB. So this is a bug in the code in method configure() of class: BenchData inside class RawErasureCoderBenchmark of test environment itself.
To fix the same we added a precondition to check argument like of chunkSizeKB.
After adding the precondition the test failed for all -ve values of dataSizeMB with Illegal argument exception.

Added assume statement in the test code to check for only +ve of numThread, chunkSizeKB and dataSizeMB. After this all our valuesets passes with ignoring tests of bad values.