Parameterizing the whole class
Found 2 globally declared and instantiated variables => {source -> stores the object to be serialized, serDeser -> stores an instance of json serializer} which can be parameterized.

As no other potential parameters were found and at least one of these two global variable is used in all the tests, hence, we can parameterise the whole class with these two variables as parameters.
-> Changed the class provided in RunWith annotation (JUnitParamsRunner.class -> Parameterized.class)
-> Changed the declaration of the two global variables, removed final keyword and made them public. And made them parameters by adding @Parameterized.Parameter annotation.
-> Removed the instantiation of the two variables and instead added their default values as the 1st value-set under the method testData in @Parameterized.Parameters annotation.
-> All the tests ran successfully.