Test 2 :- testDUGetUsedWillNotReturnNegative
Found one potential parameter - the input to the funciton incDfsUsed in line 154, which is the value to increment the cached value of used space.
Passed a variable to function, and added its default value as 1st valueset in a method valueSetForDfsUsedValue linked to test via @Parameters annotation.
The test passed.