found & fixed flaky behaviour with added value set
added 1 value set and found flaky behaviour
While the thread.sleep is called the test expects the DU to get refreshed to verify the used space in the last assertion in line 188.
The refresh time of Du isn't constant but is randomly selected between interval-jitter and interval+jitter. As in our case interval + jitter is greater than sleep time, hence, Du didn't refresh for some random scenarios and the assertion failed.
 Fixed the same by adding a condition to check the same in an assume statement.
 The test passes successfully.