added more value sets & fixed
out of 8 added value sets 2 failed
value set 7 failed as fileSize is -ve which is used to create an array in createFile method.
value set 8 failed -> assert statement on line 189 failed -> as initialUse value was provided negative. When initialUse is -ve, as per the written logic, the DU is refreshes and the actual usedSpace value is picked. Hence initial value is not honoured in this case.