more value sets and fix
added 2 more value sets with fileSize not a multiple of block size and they failed the test in the last assertion.
-> this didn;t happen in Test 1 as it only asserts inequality not exact values.
As per the stated assumption starting from line 92 of code, for only created files of sizes that are multiples of the block size will the used space be exactly equal to file size.
Added another assumption to check the same.
The test passed for all value-sets.