Test 2 :- testShouldRemoveElement
Found potential parameter in test 2 to be number of elements to be removed. Test 2 can be combined with test 1 as there we are already adding parameterized number of elements which can be removed and asserted.
Combined test 2's assertions  in test 1's code by adding a for loop at last to remove all elements and assert them. Also added an assertion at last to check if the collection has become empty.
The test passed for all value sets.
removed test 2's code.
Also updated comments of scenario for tests.