added value set & fixed test
added a value set in which all 3 names were equal
The test failed with a NullPointerException at newToken.getUser().getRealUser().getUserName() in the assert statement of line 253 old.
The reason for failure was, in class AbstractDelegationTokenIdentifier in method getUser, when real user is equal to owner or real user is null / empty string -> then real user's information is not appended in user info by creating a proxy user -> in line 83.
Hence to fix this in test code we added the same condition and if the condition is true we will check real user directly.
The test passed after this fix.