1 more value sets & fix
added another value set with blank values for all 3 user names,
The test failed with NullPointerException at newToken.getUser().getUserName() in assert statment of line 252 old.
The reason for failed was again a condition in class AbstractDelegationTokenIdentifier on line 73 -> if owner is null or an empty string then getUser would return null.
Hence in our case as getUser was null, calling getUserName() for a null value gave exception.
To fix the same, added the same condition in test code as well, the test passed after that.