potential bug ?
added 7 more value sets out of which last 5 failed
value sets at line 240 and 241 failed in assertions 282 and 288 respectively -> reason of failure was in KerberosName class in getShortName() method => as @ was provided in string reaml was eaqual to values following @ (hostname was null). On line 422 old, the for loop ilerates over rules to apply which were null which caused a null pointer exception to be thrown.
To pass the same we added a null check for rules in the KerberosName class, returning serviceName. Which passed the two value sets.
For last 3 value sets -> they failed with IllegalArgumentException thrown from 122 line of KerberosName Class. When name doesn't match the provided pattern and names still contains @ the exception is thrown.
To handle this we added assume statements with the same condition.
After which all tests passed.