Test 12 :- testInnerCauseNotInFormat
reused the method to provide valuesets and paramaterized the cause, the hard coded staring.
After replacing all the occurrences of hardcoded value with causeStr, the test failed for empty string as for that the false assertion passed as empty string can be found in any string. Fixed the same by adding assume statment.
Passed for all.