Test 1 & Test 2
Found test 1 and 2 to be the same only for a different integer value passed to testRead function.
Parameterized the 1st test with int bufLen and passed the param to testRead instead of the constant value.
Added default values from both the tests as 1st & 2nd value sets.
Both passed.
Removed the 2nd test.