added new value sets and fixed the test
Added 3 new value sets and 2 out of them failed
The failure were the -ve values.
Found the reason for failure to be in method testRead() of same class. The assertEquals on line 71 fails as buf only has 4 bytes, cause of below reason.
If the value of befLen is <= 0 in that case nothing is written in the ByteArrayOutputStream’s object bytesOut. But when blockCompressorStream is created and close with bytesOut as an OutputStream, the method rawWriteInt() of class BlockCompressorStream writes 4 pre-defined bytes in bytesOut.
And as buf, the byte array, is initiated with bytesOut, it would only have those 4 pre-defined bytes.
Hence, assert is not generalised for -ve values.
We made it generalised by taking max of 0 and bufLen.
All value sets passed after the fix.