added 3 new value sets & fixed
added 3 new value sets and the last one failed for tests testGetBytes and testIO with assertEquals failures in both (new line 94 & 126).
On investigation found the reason to be in the called method writeString() of class UTF8.java, which has a check if length of string is too large -> truncates string up to 0xffff/3 (65535 / 3  = 21845) if the string is any larger than this.
Fixed this by adding if statements if continue to skip that iteration if the value of string exceeds the provided length.
Using Assume would skip the test and hence skipping iterations which could be valid.
Test Passed.