added more value sets & fixed
added 5 value sets out of which 2 failed
nextInt() doesn't accept -ve or 0 value and throws illegal argument exception for them.
The 5th value set on line 63 didn't throw this exception as iterations was -ve and hence nextInt was never called.
Fixed the issue by assume statement in the getTestString method
Passed for all.