Test 1 :- testWithWritable
Found 3 potential params -> int iterations, int bound & char[] characters
passed all identified params to the test method.
replaced occurrences of hard coded values 10 & 40 with variables iterations and bound respectively.
removed the global declaration and initialization of alphabet, the character array, as it is not being used anywhere in the class and replaced its use in test method with the passed parameter characters.
Added the first value set as the default values of all variables.
Passed.