bug confirmed and fixed
Found a small bug in class DefaultStringifier in the main code
In the method storeArray() 1st element of array items i.e. items[0] is used to get the class for stringifier (old line 162) without checking if the size of array is greater than 0.
We added a check to fix the bug -> if array length is 0 then there is no need to make the stringifier as the loop on line 165 would not run.
Shifted the code to create and initialize the object of StringBuilder and used it in the added check to set the configuration -> provided keyName to empty array and return.
After the fix the test passed for all the value sets.