Test 3 :- apply
Couldn't find anything to paramaterize in test 1 and 2.
Found hard coded value of depth as -1 as potential param.
Used refactor -> introduce parameter to pass it as a param to method.
Added its default value as 1st value set.
Passed.
Also thought of parameterizing 2 (number of times apply is invoked) but couldn't figure out how to generalize when method of mokito for the same.