Added more value sets and assume to fix
 all values of depth failed except for -1.
On investigation found the reason in class FilterExpression, in method apply() -> if expression != null, which isn't null in our case as we have called the constructor with expr on line 49, then expression.apply is called with depth = -1.
Hence only depth's value as -1 is accepted.
Added assume statment to ensure the same.