Test 6 :- applyGlobNotMatch
used refactor -> introduce parameter on n*e and notmatch. Was able to change the name of notMatch as it was a substring so intelliJ gave a pop up window to set name, but not for n*e, it automatically named it to variable arg.
Added default value as 1st value set
passed.