Test 2 :- testAddToEmptyLinkedList
Found test 2 to exactly similar to test 1 except for creating a Linked List instead of an arrayList.
Found constant string record1 as potential param
instead of just 1 string we can add multiple strings to the declared list of Strings.
 Hence, passed a param of String array [] stringArr -> whose values are added one after one just like in the test using .add method.
 generalized the assertions to check from the input param and added 2 for loops to first add the elements and then assert.
Reused the created method to pass string array in test 1, by renaming it for a better fit.
All value sets passed.