Test 3 :- testVarArgArrayLists
Found 2 params as String... elements to be passed while creating new array list and int nMore -> number of elements to be added afterwards.
Passed the two params to test.
replaced the static 3 record strings with variable elements
replaced list.add with a for loop iterating nMore times and adding elements to the array.
generalized assertions by adding 2 loops (for and while) to check all the added elements.
Added a method to provide value sets and passed the default values as 1st value set.
passed.