Test 4:- testItrArrayLists
Found potential params int nMore, String... elements like in the test above.
Reused the same method as test 3 to provide value sets.
generalized the test method by adding for loops.
in line 124 used set.size instead of elements,size as we are adding set's elements to list, hence it would not have redundant values which might be present in elements.
passed for all .