Test 2 :- testPathResolveToLinkNotResolveLastComponent
Found 4 potential params in test 2
the first 3 are same as test 1 and the 4th is URI expectedURI, as they changed expectedURI in their 2nd test snippit
Found that the only difference between test 1 and 2 is the value of resolveLastComponent -> hence decided to merge the two.
To merge as we'll need to provide complex object as param (expectedURI) changed the value set providing method of test 1 to method.
All tests of test 1 passed successfully.