Test 5, 6 & 7
these tests were already generalized by intelliJ hence added their inputs to call test 1 as different value sets for test 1.
failed for test 5 and 7
the reason for failure was the extra verification detected on line 87
when we added the condition to verify sencond only when 1st is pass -> we missed the condition for stop
added the same
all passed
removed the code for tests 5, 6 and 7.