Test 8, 9 & 10
found test 8, 9 and 10 to be similar apart from calling differnt functions on child
added a switch statement to generalize the test
each case of switch has the calls to different functions as per different tests
before and after the switch there are common calls which were shared by all three functions.
Added names of different functions which are called as switch cases and also as value sets.
all passed
deleted code for tests 9 and 10.