added more value sets & assume
added more value sets, out of which the last one failed with java.lang.OutOfMemoryError: Java heap space as the writeJunk method keeps writing in the DataOutputBuffer.
to fix the same added an assume statement to ignore tests where the iterations are more than square of the default value.
all passed.