Test 6 & 7
for these already generalized tests -> added default value of test 6 and input values to test 6's method (for test 7) as value sets for test 6.
Along with proper comments.
Also update the name of test 6 to be more generalized.
All value sets passed.
removed the code of test 7 from the file.