new Test 3 fix
the newly added value set for new test 3 actually fails
Found the reason to be the value set class FailingStopInStartService doesn't have fail neither in init nor in start which the given test tests.
To fix this added an assume statement to check the exit code and only allow for -1 and -2 which are exit codes of  pre defined class with either fail in init or in start.
All tests in this class passes after this.